Add Toolbar component with text formatting and export functionality
